Create SMBmap_install_validation_removal
This program will create a new directory on the local C drive, connect and authenticate to the local Samba server, copy the folder containing all the required dependencies from that server onto the newly created directory (local) , then invoke installer/install executable to install it. 

Once installed it will check for the proper file path to validate that it has properly installed. After confirming this it will remove the folder containing the install files (staging branch) due to the large size of them (spartan was close to 100gb so were a few of the other programs I had written more scripts to install in a similar fashion (MatLab for example) 

Then it will sleep and set the execution policy back to restricted and exit